Transcription/requirements.txt
flask
vosk
sounddevice
scipy
numpy
pydub
noisereduce
transformers
spacy
torch
torchaudio
librosa
soundfile
faster-whisper
